+++ game/sound/989snd/locator.h
@@ -13,7 +13,7 @@ class locator {
   virtual ~locator() = default;
   virtual SoundBank* get_bank_by_handle(u32 id) = 0;
   virtual MusicBank* get_bank_by_name(u32 id) = 0;
-  virtual u16* get_bank_samples(u32 id) = 0;
+  virtual u8* get_bank_samples(u32 id) = 0;
   virtual MIDIBlock* get_midi(u32 id) = 0;
 };
 }  // namespace snd
